agents/spring-voyage-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spring-voyage-agent"
version = "0.1.0"
description = "Spring Voyage Agent SDK — Python lifecycle-hooks package (initialize / on_message / on_shutdown)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    # Upper-pinned to <1 because a2a-sdk 1.0.0 (April 2026) reorganised its
    # package layout: `a2a.server.apps` and `a2a.utils.artifact` are gone.
    # Lift the bound when imports are updated to the 1.x surface.
    "a2a-sdk[http-server]>=0.3.20,<1",
    "uvicorn>=0.30,<1",
    # PyYAML for loading agent-definition.yaml (optional at runtime;
    # the SDK falls back to JSON if YAML is absent).
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
spring-voyage-agent = "spring_voyage_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["spring_voyage_agent"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
